requirements.txt
drissionpage>=4.0.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai>=1.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pypdf>=4.0.0
